pyproject.toml
[project]
name = "image-tagger-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.13",
    "langchain-community>=0.3.13",
    "langchain-core>=0.3.28",
    "langchain-openai>=0.2.14",
    "openai>=1.58.1",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]
